+++ packages/x-components/src/x-modules/history-queries/components/history-queries.vue
@@ -70,7 +70,7 @@
      *
      * @internal
      */
-    @Getter('historyQueries', 'historyQueries')
+    @Getter('historyQueries', 'historyQueriesWithResults')
     public historyQueries!: HistoryQueryModel[];
   }
 </script>