+++ packages/x-svg-converter/.eslintrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+  parserOptions: {
+    tsconfigRootDir: __dirname,
+    project: 'tsconfig.eslint.json'
+  }
+};